% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{add_grossnet_column}
\alias{add_grossnet_column}
\title{Add a Gross.Net column and delete matrices}
\usage{
add_grossnet_column(
  .agg_data,
  countries,
  years,
  gross_net = Recca::efficiency_cols$gross_net,
  R = "R",
  U = "U",
  U_feed = "U_feed",
  U_eiou = "U_EIOU",
  r_eiou = "r_EIOU",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
  ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
  ex_fd = Recca::aggregate_cols$aggregate_demand
)
}
\arguments{
\item{.agg_data}{Input data containing gross and net final demand aggregate columns.}

\item{countries}{The countries to work on.}

\item{years}{The years to work on.}

\item{gross_net}{The name of the column that identifies gross or net final demand aggregates.
Default is \code{Recca::efficiency_cols$gross_net}.}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y, S_units}{Columns of matrices to be deleted.}

\item{gross, net}{Strings inserted into the \code{grossnet} column. See \code{Recca::efficiency_cols}.}

\item{ex_fd_gross, ex_fd_net, ex_fd}{Names for columns in \code{.agg_data} for gross and net final demand aggregations.}
}
\value{
A data frame with matrices removed and a new \code{GrossNet} column.
}
\description{
Between calculating the gross and net final demand aggregates
and calculating efficiencies,
it is helpful to add a step wherein
delete the matrices (lots of data!) and add a \code{GrossNet} column.
This function performs those tasks.
}
