% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_df_cols}
\alias{aggregation_df_cols}
\title{Aggregation data frame column names}
\format{
A string list with 5 entries.
\describe{
\item{product_aggregation}{The name of the metadata column that tells about product aggregation. "Product.Aggregation"}
\item{industry_aggregation}{The name of the metadata column that tells about industry aggregation. "Industry.Aggregation"}
\item{specified}{The value that indicates products or industries remain is specified. "Specified"}
\item{despecified}{The value that indicates products or industries have been despecified and aggregated. "Despecified"}
\item{industry_aggregation}{The value that indicates products or industries have been grouped. "Grouped"}
}
}
\usage{
aggregation_df_cols
}
\description{
A string list containing names of column names and values for aggregatino data frames.
}
\examples{
aggregation_df_cols
}
\keyword{datasets}
