% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_file_cols}
\alias{aggregation_file_cols}
\title{Aggregation table column names}
\format{
A string list with 2 entries.
\describe{
\item{many_colname}{The name of the many column in each tab.}
\item{few_colname}{The name of the few column in each tab.}
}
}
\usage{
aggregation_file_cols
}
\description{
A string list containing names of column names in aggregation tables.
}
\examples{
aggregation_file_cols
}
\keyword{datasets}
