% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_file_tab_names}
\alias{aggregation_file_tab_names}
\title{Aggregation file tab information}
\format{
A string list with 5 entries.
\describe{
\item{continent_aggregation}{The name of the continent aggregation tab, "continent_aggregation".}
\item{world_aggregation}{The string name of the world aggregation tab. "world_aggregation".}
\item{ef_product_aggregation}{The string name of the final energy product aggregation tab. "ef_product_aggregation".}
\item{eu_product_aggregation}{The string name of the useful energy product aggregation tab. "eu_product_aggregation".}
\item{ef_sector_aggregation}{The string name of the final energy sector aggregation tab. "ef_sector_aggregation".}
}
}
\usage{
aggregation_file_tab_names
}
\description{
A string list containing the aggregation file's tab names.
}
\examples{
aggregation_file_tab_names
}
\keyword{datasets}
