% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calc_efficiencies}
\alias{calc_efficiencies}
\title{Calculate efficiencies from an aggregates data frame}
\usage{
calc_efficiencies(
  .aggregates,
  stage_colname = PFUAggDatabase::sea_cols$stage_colname,
  ex_colname = PFUAggDatabase::sea_cols$ex_colname,
  eta_pf_colname = PFUAggDatabase::efficiency_cols$eta_pf,
  eta_fu_colname = PFUAggDatabase::efficiency_cols$eta_fu,
  eta_pu_colname = PFUAggDatabase::efficiency_cols$eta_pu,
  primary = IEATools::all_stages$primary,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  gross_net_colname = PFUAggDatabase::sea_cols$gross_net_colname,
  gross = PFUAggDatabase::gross_net_metadata$gross,
  net = PFUAggDatabase::gross_net_metadata$net
)
}
\arguments{
\item{.aggregates}{The data frame from which efficiencies are to be calculated.
This data frame should be the output of the}
}
\value{

}
\description{
Calculate efficiencies from an aggregates data frame
}
