% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_effects_aggregations.R
\name{calculate_effects_aggregations}
\alias{calculate_effects_aggregations}
\title{Calculate effects aggregates}
\usage{
calculate_effects_aggregations(
  .psut_data,
  countries,
  years,
  p_industries,
  fd_sectors,
  method = "SVD",
  tol_invert = .Machine$double.eps
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices. It should be wide by matrices.}

\item{countries}{The countries to analyze.}

\item{years}{The years to analyze.}

\item{p_industries}{Industries that count for primary energy aggregates.}

\item{fd_sectors}{Final demand sectors that count for final demand aggregates.}

\item{method}{Tells how to invert matrices. Default is "SVD". See details.}

\item{tol}{The tolerance for detecting linear dependencies in the columns of matrices to be inverted.
Default is \code{.Machine$double.eps}.}
}
\value{
A data frame of effects aggregates.
}
\description{
This function uses \code{Recca::effects_aggregates()} internally
to calculate effects (downstream) aggregates of resource matrix energy carriers.
}
\details{
Effects aggregations involve a matrix inversion step.
The \code{method} argument specifies which method should be used for
calculating the inverse.
See \code{matsbyname::invert_byname()}.

Both \code{tol} and \code{method} should be a single values and apply to all matrices being inverted.
}
