% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_finaluseful_ex_data}
\alias{calculate_finaluseful_ex_data}
\title{Create a data frame containing final and useful aggregate energy/exergy data}
\usage{
calculate_finaluseful_ex_data(
  .sutdata,
  countries,
  years,
  fd_sectors,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  ex = PFUAggDatabase::sea_cols$ex_colname
)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}

\item{countries}{The countries for which primary energy and exergy data are to be calculated.}

\item{years}{The years for which primary energy and exergy data are to be calculated.}

\item{fd_sectors}{A character vector of final demand sectors.}

\item{country}{The name of the country column in \code{.sutdata}.
Default is \code{IEATools::iea_cols$country}.}

\item{year}{The name of the year column in \code{.sutdata}.
Default is \code{IEATools::iea_cols$year}.}

\item{ex}{The name of the energy and exergy column in \code{.sutdata}.
Default is \code{PFUAggDatabase::sea_cols$ex_colname}.}
}
\value{
A data frame containing final and useful energy/exergy values aggregated by total,
sector and product.
}
\description{
This functions creates a single data frame containing final and useful
energy/exergy by country, year, method, energy quantification,
and grouping variable (Total, Product, and Sector) using the functions:
\code{calculate_fu_ex_total}, \code{calculate_fu_ex_sector}, \code{calculate_fu_ex_product}
and binding the outputs of these functions into a single data frame.
}
\examples{
library(Recca)
finaluseful_data <- Recca::UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) \%>\%
  dplyr::mutate(Method = "PCM") \%>\%
  calculate_finaluseful_ex_data(countries = "all",
                                years = "all",
                                fd_sectors = c("Residential"))
}
