% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_aggregations.R
\name{calculate_footprint_aggregations}
\alias{calculate_footprint_aggregations}
\title{Calculate footprint aggregates}
\usage{
calculate_footprint_aggregations(
  .psut_data,
  countries,
  years,
  p_industries,
  fd_sectors
)
}
\arguments{
\item{fd_sectors}{}
}
\description{
This function uses \code{Recca::footprint_aggregates()} internally.
}
