% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_fu_ex_product}
\alias{calculate_fu_ex_product}
\title{Calculate total final consumption of final and useful energy by product}
\usage{
calculate_fu_ex_product(
  .sutdata,
  fd_sectors,
  country_colname = IEATools::iea_cols$country,
  method_colname = IEATools::iea_cols$method,
  energy_type_colname = IEATools::iea_cols$energy_type,
  last_stage_colname = IEATools::iea_cols$last_stage,
  year_colname = IEATools::iea_cols$year,
  sector_colname = PFUAggDatabase::sea_cols$sector_colname,
  fd_sectors_colname = PFUAggDatabase::sea_cols$fd_sectors_colname,
  e_product_colname = PFUAggDatabase::sea_cols$e_product_colname,
  stage_colname = PFUAggDatabase::sea_cols$stage_colname,
  gross_net_colname = PFUAggDatabase::sea_cols$gross_net_colname,
  agg_by_colname = PFUAggDatabase::sea_cols$agg_by_colname,
  ex_colname = PFUAggDatabase::sea_cols$ex_colname,
  ex_net_colname = Recca::aggregate_cols$net_aggregate_demand,
  ex_gross_colname = Recca::aggregate_cols$gross_aggregate_demand,
  net_value = PFUAggDatabase::gross_net_metadata$net_value,
  gross_value = PFUAggDatabase::gross_net_metadata$gross_value,
  all_value = PFUAggDatabase::agg_metadata$all_value,
  product_value = PFUAggDatabase::agg_metadata$product_value
)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}

\item{fd_sectors}{A character vector of final demand sectors.}

\item{country_colname, method_colname, energy_type_colname, last_stage_colname, year_colname}{See \code{IEATools::iea_cols}.}

\item{sector_colname, fd_sectors_colname, e_product_colname, stage_colname, gross_net_colname, agg_by_colname, ex_colname, ex_net_colname, ex_gross_colname}{See \code{PFUAggDatabase::sea_cols}.}

\item{net_value, gross_value}{See \code{PFUAggDatabase::gross_net_metadata}.}

\item{all_value, product_value}{See \code{PFUAggDatabase::agg_metadata}.}
}
\value{
A data frame containing aggregate final and useful energy/exergy data by product
}
\description{
Calculate the total final consumption (TFC) at the final and useful stages
(along with any additional stages) by product.
This function first uses the uses \code{create_fd_sectors_list()} function,
with a user-supplied set of final demand sectors \code{fd_sectors}
to identify the final demand sectors desired for analysis.
The \code{Recca::finaldemand_aggregates()} function is then applied to \code{.sutdata}
data frame by product, to calculate the total final consumption across
all of the sectors supplied in \code{fd_sectors} for each product.
}
\examples{
library(Recca)
tfc_product <- Recca::UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) \%>\%
  dplyr::mutate(Method = "PCM") \%>\%
  calculate_fu_ex_product(fd_sectors = c("Residential"))
}
