% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_p_ex_total}
\alias{calculate_p_ex_total}
\title{Calculate total energy supply}
\usage{
calculate_p_ex_total(
  .sutdata,
  p_industry_prefixes,
  country_colname = IEATools::iea_cols$country,
  method_colname = IEATools::iea_cols$method,
  energy_type_colname = IEATools::iea_cols$energy_type,
  year_colname = IEATools::iea_cols$year,
  flow_colname = PFUAggDatabase::sea_cols$flow_colname,
  e_product_colname = PFUAggDatabase::sea_cols$e_product_colname,
  stage_colname = PFUAggDatabase::sea_cols$stage_colname,
  gross_net_colname = PFUAggDatabase::sea_cols$gross_net_colname,
  agg_by_colname = PFUAggDatabase::sea_cols$agg_by_colname,
  p_ind_comp_colname = PFUAggDatabase::sea_cols$p_ind_comp_colname,
  p_ind_prefix_colname = PFUAggDatabase::sea_cols$p_ind_prefix_colname,
  ex_colname = PFUAggDatabase::sea_cols$ex_colname,
  ex_p_colname = PFUAggDatabase::sea_cols$ex_p_colname,
  primary_value = IEATools::all_stages$primary,
  all_value = PFUAggDatabase::agg_metadata$all_value,
  total_value = PFUAggDatabase::agg_metadata$total_value
)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT) matrices.}

\item{p_industry_prefixes}{A character vector of primary energy industry prefixes.
Usually "Resources", "Imports", and "Stock changes".}

\item{country_colname, method_colname, energy_type_colname, year_colname}{See \code{IEATools::iea_cols}.}

\item{flow_colname, e_product_colname, stage_colname, gross_net_colname, agg_by_colname, p_ind_comp_colname, p_ind_prefix_colname, ex_colname, ex_p_colname}{See \code{PFUAggDatabase::sea_cols}.}

\item{primary_value}{The string "Primary", representing the Primary stage of the energy conversion chain, see \code{IEATools::all_stages}.}

\item{all_value, total_value}{See \code{PFUAggDatabase::agg_metadata}.}
}
\value{
A data frame containing aggregate primary energy/exergy data by total (total energy supply (TES)).
}
\description{
Calculate the total energy supply (TES) in primary energy or exergy terms.
This metric was formerly called the total primary energy supply (TPES) by the IEA.
This function first uses the uses \code{Recca::find_p_industry_names()} function,
with a user-supplied set of primary industry prefixes \code{p_industry_prefixes}
to identify the primary industries desired for analysis.
The \code{Recca::primary_aggregates()} function is then applied to the \code{.sutdata} data frame by total,
to calculate the total energy supply across all products and flows.
}
\examples{
library(Recca)
total_energy_supply <- Recca::UKEnergy2000mats \%>\%
tidyr::pivot_wider(names_from = matrix.name,
                   values_from = matrix) \%>\%
  dplyr::mutate(Method = "PCM") \%>\%
  calculate_p_ex_total(p_industry_prefixes = list(c("Resources", "Imports")))
}
