% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_pfu_aggregates}
\alias{calculate_pfu_aggregates}
\title{Calculate primary, final, and useful aggregates}
\usage{
calculate_pfu_aggregates(
  .agg_data,
  countries,
  years,
  gross_net = Recca::efficiency_cols$gross_net,
  R = "R",
  U = "U",
  U_feed = "U_feed",
  U_eiou = "U_EIOU",
  r_eiou = "r_EIOU",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  last_stage = Recca::psut_cols$last_stage,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_f = IEATools::aggregate_cols$aggregate_final,
  ex_u = IEATools::aggregate_cols$aggregate_useful,
  ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
  ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
  ex_fd = Recca::aggregate_cols$aggregate_demand
)
}
\arguments{
\item{.agg_data}{Input data containing gross and net final demand aggregate columns.}

\item{countries}{The countries to work on.}

\item{years}{The years to work on.}

\item{gross_net}{The name of the column that identifies gross or net final demand aggregates.
Default is \code{Recca::efficiency_cols$gross_net}.}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y, S_units}{Columns of matrices to be deleted.}

\item{gross, net}{Strings inserted into the \code{grossnet} column. See \code{Recca::efficiency_cols}.}

\item{last_stage}{The string name of a column containing last stages of energy conversion chains.
Default is \code{Recca::psut_cols$last_stage}.}

\item{final, useful}{String names of last stages. Defaults from \code{IEATools::all_stages}.}

\item{ex_p, ex_f, ex_u}{Names for columns in \code{.agg_data} for primary, final, and useful aggregations.}

\item{ex_fd, ex_fd_gross, ex_fd_net}{Names for columns of total, gross, and net final demand aggregates.
Defaults from \code{Recca::aggregate_cols}.}
}
\value{
A data frame with matrices removed and a new \code{GrossNet} column.
}
\description{
In addition to calculating primary, final, and useful aggregates,
this function removes matrix columns and adds a \code{GrossNet} column to the data frame.
}
