% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calculate_pfu_efficiencies}
\alias{calculate_pfu_efficiencies}
\title{Pivot data frame to calculate PFU efficiencies}
\usage{
calculate_pfu_efficiencies(
  .eta_pfu_data,
  countries,
  years,
  ex_p = IEATools::aggregate_cols$aggregate_primary,
  ex_f = IEATools::aggregate_cols$aggregate_final,
  ex_u = IEATools::aggregate_cols$aggregate_useful,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu
)
}
\arguments{
\item{.eta_pfu_data}{A data frame produced by \code{add_grossnet_column()}.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{ex_p, ex_f, ex_u}{Names of primary, final, and useful energy and exergy aggregate columns (respectively).
Defaults are from \code{IEATools::aggregate_cols}.}

\item{eta_pf, eta_fu, eta_pu}{The names of efficiency columns: primary-to-final, final-to-useful, and primary-to-useful, respectively.
Defaults from \code{Recca::efficiency_cols}.}
}
\value{
A data frame with ECC stage efficiencies.
}
\description{
This function pivots the data frame produced by \code{add_grossnet_column()}
to obtain primary-final, final-useful, and primary-useful efficiencies.
}
