% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_aggregations.R
\name{calculate_sector_aggregates}
\alias{calculate_sector_aggregates}
\title{Calculate final demand sector sums}
\usage{
calculate_sector_aggregates(
  .psut_data,
  countries,
  years,
  fd_sectors,
  pattern_type = "leading",
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  Y = Recca::psut_cols$Y,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{fd_sectors}{The sectors that count for final demand.}

\item{pattern_type}{Where to look for sectors.
Default is "leading".}

\item{U}{The name of the use matrix column in \code{.psut_mats}.
Default is \code{Recca::psut_cols$U}.}

\item{U_feed}{The name of the feed matrix column in \code{.psut_mats}.
Default is \code{Recca::psut_cols$U_feed}.}

\item{Y}{The name of the final demand matrix column in \code{.psut_mats}.
Default is \code{Recca::psut_cols$Y}.}

\item{net_aggregate_demand}{The name of the net aggregate demand column on output.
Default is \code{Recca::aggregate_cols$net_aggregate_demand}.}

\item{gross_aggregate_demand}{The name of the gross aggregate demand column on output.
Default is \code{Recca::aggregate_cols$gross_aggregate_demand}.}
}
\value{
A data frame of sector energy and exergy sums.
}
\description{
This function calculates final demand aggregates by sector.
}
