% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_aggregations.R
\name{calculate_sector_fu_agg_eta}
\alias{calculate_sector_fu_agg_eta}
\title{Calculate final and useful sector sums and efficiencies}
\usage{
calculate_sector_fu_agg_eta(
  .psut_data,
  countries,
  years,
  fd_sectors,
  pattern_type = "leading",
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  year = Recca::psut_cols$year,
  last_stage = Recca::psut_cols$last_stage,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  aggregate_demand = Recca::aggregate_cols$aggregate_demand,
  gross_net = Recca::efficiency_cols$gross_net,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  sector = "Sector",
  colnames = "colnames",
  rowtypes = "rowtypes",
  coltypes = "coltypes",
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  eta_fu = Recca::efficiency_cols$eta_fu
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{fd_sectors}{The sectors that count for final demand.}

\item{pattern_type}{Where to look for sectors.
Default is "leading".}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y, S_units}{The names of the matrix columns in \code{.psut_mats}.
Defaults from \code{Recca::psut_cols}.}

\item{year}{The string name of the year column.
Default is \code{Recca::psut_cols$year}.}

\item{last_stage}{The string name of the last stage column.
Default is \code{Recca::psut_cols$last_stage}.}

\item{aggregate_demand, gross_aggregate_demand, net_aggregate_demand}{The names of columns of aggregate demand.
Defaults from \code{Recca::aggregate_cols}.}

\item{gross_net}{The name of the column that tells whether aggregates are gross or net.
Default is \code{Recca::efficiency_cols$gross_net}.}

\item{gross, net}{Strings identifying gross or net quantities.
Defaults from \code{Recca::efficiency_cols}.}

\item{sector}{The name of the sector column.
Default is "Sector".}

\item{colnames, rowtypes, coltypes}{Names of columns deleted after expanding matrices.
Defaults are "colnames", "rowtypes", and "coltypes", respectively.}

\item{final, useful}{The string names for final and useful aggregates.
Defaults from \code{IEATools::all_stages}.}

\item{eta_fu}{The string name of the final-to-useful efficiency.
Default is \code{Recca::efficiency_cols$eta_fu}.}
}
\value{
A data frame of sector energy and exergy sums.
}
\description{
This function calculates final demand aggregates by sector
for both final and useful stages and resulting efficiencies.
}
