% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopRY.R
\name{chop_Y_eccs}
\alias{chop_Y_eccs}
\title{Chop the ECC for \strong{Y}}
\usage{
chop_Y_eccs(
  .psut_data,
  countries,
  years,
  method = "SVD",
  tol = .Machine$double.eps
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices. It should be wide by matrices.}

\item{countries}{The countries to analyze.}

\item{years}{The years to analyze.}

\item{method}{Tells how to invert matrices. Default is "SVD". See details.}

\item{tol}{The tolerance for detecting linear dependencies in the columns of matrices to be inverted.
Default is \code{.Machine$double.eps}.}

\item{p_industries}{Industries that count for primary energy aggregates.}

\item{fd_sectors}{Final demand sectors that count for final demand aggregates.}
}
\value{
A data frame of chopped final demand matrix ECCs.
}
\description{
This function uses \code{Recca::chop_Y()} internally
to calculate new ECCs for each row or column in the \strong{Y} matrix.
}
\details{
Chopping in the \strong{Y} matrix and calculating a new ECC
involves a matrix inversion step.
The \code{method} argument specifies which method should be used for
calculating the inverse.
See \code{matsbyname::invert_byname()}.

Both \code{tol} and \code{method} should be a single values and apply to all matrices being inverted.
}
