% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopRY.R
\name{chopRY}
\alias{chopRY}
\alias{chop_R_eccs}
\alias{chop_Y_eccs}
\title{Chop the \strong{R} and \strong{Y} matrices}
\usage{
chop_R_eccs(
  .psut_data,
  countries,
  years,
  do_chops,
  method = "SVD",
  tol_invert = .Machine$double.eps
)

chop_Y_eccs(
  .psut_data,
  countries,
  years,
  do_chops,
  method = "SVD",
  tol_invert = .Machine$double.eps
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices. It should be wide by matrices.}

\item{countries}{The countries to analyze.}

\item{years}{The years to analyze.}

\item{do_chops}{A boolean that tells whether to do the \strong{R} and \strong{Y} chops.}

\item{method}{Tells how to invert matrices. Default is "SVD". See details.}

\item{tol_invert}{The tolerance for detecting linear dependencies in the columns of matrices to be inverted.
Default is \code{.Machine$double.eps}.}
}
\value{
A data frame of chopped \strong{R} or \strong{Y} matrix ECCs.
}
\description{
These functions use \code{Recca::chop_R()} and \code{Recca::chop_Y()} internally
to calculate new ECCs for each column in the \strong{R} matrix
and for each row and column for the \strong{Y} matrix.
}
\details{
Chopping in the \strong{R} and \strong{Y} matrices and calculating a new ECC
involves a matrix inversion step.
The \code{method} argument specifies which method should be used for
calculating the inverse.
See \code{matsbyname::invert_byname()} for additional details.

Both \code{tol_invert} and \code{method} should be a single values and apply to all matrices being inverted.

When \code{do_chops} is \code{FALSE}, a data frame is returned
with no rows and the same columns as \code{.psut_data}.
}
