% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_aggregations.R
\name{continent_aggregation}
\alias{continent_aggregation}
\title{Calculate continent aggregation}
\usage{
continent_aggregation(
  PSUT,
  continents,
  continent = "Continent",
  many_colname = IEATools::iea_cols$country,
  few_colname = continent
)
}
\arguments{
\item{PSUT}{A data frame of PSUT matrices}

\item{continents}{The continents for which aggregation is desired.}

\item{continent}{The name of the continent column in the \code{PSUT} data frame.
Default is "Continent".}

\item{many_colname}{The name of the column of many things that will be aggregated into continents.
Default is \code{IEATools::iea_cols$country}.}

\item{few_colname}{The name of the column of few things that will remain on output.
Default is the same as the value of the \code{continent} argument.}
}
\value{
A data frame in which countries are aggregated to continents according to
the aggregation map.
}
\description{
This function is a wrapper for \code{Recca::region_aggregates()}
that also filters on \code{continents},
thereby enabling parallel processing across continents.
}
