% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pipelines.R
\name{create_continent_table}
\alias{create_continent_table}
\title{Load and create a data frame for use with continent aggregation}
\usage{
create_continent_table(
  exemplar_table_path,
  exemplar_table_tab = "exemplar_table",
  region_code = "Region.code",
  continent = "Continent",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{exemplar_table_path}{The path to the exemplar table}

\item{exemplar_table_tab}{The name of the tab in the file at \code{exemplar_table_path}
where the region information exists.
Default is "exemplar_table".}

\item{region_code}{The name of a column containing region codes.
Default is "Region.code".}

\item{country}{The name of the country column in the outgoing data frame.
Default is \code{IEATools::iea_cols$country}.}

\item{year}{The name of the year column in the outgoing data frame.
Default is \code{IEATools::iea_cols$year}.}
}
\value{
A data frame with columns "Country", "Year", and \code{region_code}.
}
\description{
The exemplar table is assumed to be an Excel file with the following columns:
\code{region_code} and years (as numbers).
The body of the table should contain 3-letter codes
of countries.
The exemplar table is assumed to be on the \code{exemplar_table_tab} tab of the Excel file.
}
