% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_aggregations.R
\name{create_fd_sectors_list}
\alias{create_fd_sectors_list}
\title{Compile the list of final demand sectors}
\usage{
create_fd_sectors_list(iea_fd_sectors, sector_aggregation_map)
}
\arguments{
\item{iea_fd_sectors}{The list of final demand sectors, according to the IEA.}

\item{sector_aggregation_map}{The aggregation map used to aggregate final demand sectors.}
}
\value{
A larger list of final demand sectors comprised of
the items in \code{iea_fd_sectors} and the names of \code{sector_aggregation_map}.
}
\description{
The \code{IEATools} package contains a list of final demand sectors
in \code{IEATools::fd_sectors}.
However, we aggregate and rename sectors according to an aggregation map.
The names of the aggregation map items should also count as
final demand sectors.
This function adds the aggregated final demand sector names
to the \code{IEATools} final demand sectors.
}
