% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficiency_cols}
\alias{efficiency_cols}
\title{Efficiency columns}
\format{
A string list with 3 entries.
\describe{
\item{eta_pf}{The string indicating primary-to-final efficiency. "eta_pf".}
\item{eta_fu}{The string indicating final-to-useful efficiency. "eta_fu".}
\item{eta_pu}{The string indicating primary-to-final efficiency. "eta_pu".}
}
}
\usage{
efficiency_cols
}
\description{
A string list containing names of efficiency columns.
}
\examples{
efficiency_cols
}
\keyword{datasets}
