% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{efficiency_pipeline}
\alias{efficiency_pipeline}
\title{Calculate PFU aggregations and efficiencies}
\usage{
efficiency_pipeline(
  .psut_data,
  p_industries,
  fd_sectors,
  piece = "noun",
  notation = list(RCLabels::bracket_notation, RCLabels::arrow_notation),
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  country = Recca::psut_cols$country,
  year = Recca::psut_cols$year,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  gross_net = Recca::efficiency_cols$gross_net,
  last_stage = Recca::psut_cols$last_stage,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_f = IEATools::aggregate_cols$aggregate_final,
  ex_u = IEATools::aggregate_cols$aggregate_useful,
  ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
  ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT data or a slice (row) of the data frame.}

\item{p_industries}{A string vector of primary industries.}

\item{fd_sectors}{A string vector of final demand sectors.}

\item{piece}{The piece of the labels used for matching.
Default is "noun".}

\item{notation}{The notation used for row and column labels.
Default is \code{list(RCLabels::bracket_notation, RCLabels::arrow_notation)}.}

\item{pattern_type}{The pattern type to be used for row and column matching.
Default is "exact".}

\item{prepositions}{A list of prepositions for row and column labels.
Default is \code{RCLabels::prepositions_list}.}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed, S_units, country, year, last_stage}{See \code{Recca::psut_cols}.}

\item{gross, net, gross_net}{See \code{Recca::efficiency_cols}.}

\item{final, useful}{See \code{IEATools::all_stages}.}

\item{ex_p, ex_fd_gross, ex_fd_net, ex_fd}{See \code{Recca::aggregate_cols}.}

\item{ex_f, ex_u}{See \code{IEATools::aggregate_cols}.}

\item{eta_pf, eta_fu, eta_pu}{See \code{Recca::efficiency_cols}.}
}
\value{
A data frame of metadata columns and efficiencies
}
\description{
Calculate PFU aggregations and efficiencies
}
