% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{exiobase_coeffs}
\alias{exiobase_coeffs}
\title{Calculate Exiobase coefficients}
\usage{
exiobase_coeffs(phi_vecs)
}
\arguments{
\item{phi_vecs}{A data frame of phi (exergy-to-energy ratio) coefficients.}
}
\value{
A data frame of Exiobase coefficients.
}
\description{
The CL-PFU Database supplies coefficients
to Exiobase.
This function calculates those coefficients
and returns in a data frame.
}
