% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{filter_countries}
\alias{filter_countries}
\title{Filter countries}
\usage{
filter_countries(
  .psut_data,
  countries = "all",
  country = IEATools::iea_cols$country
)
}
\arguments{
\item{.psut_data}{The entire PSUT data frame, read from the pinboard.}

\item{countries}{A vector of 3-letter country codes for the countries to be used in this analysis.
Default value is "all", meaning that all countries in \code{.psut_data} will be included.}

\item{country}{The name of the country column in \code{.psut_data}.
Default is \code{IEATools::iea_cols$country}.}
}
\value{
A filtered version of \code{.psut_data}..
}
\description{
Chooses countries to use in the rest of the pipeline.
}
