% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_years.R
\name{filter_countries_and_years}
\alias{filter_countries_and_years}
\title{Filter countries}
\usage{
filter_countries_and_years(
  .psut_data,
  countries = "all",
  years = "all",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{.psut_data}{The entire PSUT data frame, read from the pinboard.}

\item{countries}{A vector of 3-letter country codes for the countries to be used in this analysis.
Default value is "all", meaning that all countries in \code{.psut_data} will be included.}

\item{years}{A numeric vector of years to be used in this analysis.
Default value is "all", meaning that ll years in \code{.psut_dat} will be included.}

\item{country, year}{See \code{IEATools::iea_cols}.}
}
\value{
A filtered version of \code{.psut_data}.
}
\description{
Chooses countries to use in the rest of the pipeline.
}
