% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_aggregations.R
\name{footprint_aggregations}
\alias{footprint_aggregations}
\title{Calculate footprint aggregates}
\usage{
footprint_aggregations(.psut_data, countries, years, p_industries, fd_sectors)
}
\arguments{
\item{fd_sectors}{}
}
\description{
This function uses \code{Recca::footprint_aggregates()} internally.
}
