% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_subpipelines.R
\name{get_one_middle_pipeline}
\alias{get_one_middle_pipeline}
\title{A pipeline for one PSUT release}
\usage{
get_one_middle_pipeline(pr, aggregation_maps_tar_str, continents_tar_str)
}
\arguments{
\item{pr}{The \code{PFUDatabase} pipeline release for which this pipeline
segment will be constructed.}

\item{aggregation_maps_tar_str}{The name of the aggregation maps target (as a string).}

\item{continents_tar_str}{The name of the continents target (as a string).}
}
\value{
A list of targets for the incoming PSUT release (\code{pr}).
}
\description{
\code{PFUAggDatabase} works with any number of PSUT releases from \code{PFUDatabase}.
This function creates a portion of a \code{targets} pipeline
for \emph{one} of those releases.
}
\details{
Some actions of this function assume target names
created by \code{setup_targets()}.
This function and \code{setup_targets()} should be considered
as a pair of functions that work together.
}
