% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gross_net_metadata}
\alias{gross_net_metadata}
\title{Gross and net metadata}
\format{
A string list with 2 entries.
\describe{
\item{gross_value}{The string indicating gross aggregation. "Gross".}
\item{net_value}{The string indicating net aggregation. "Net".}
}
}
\usage{
gross_net_metadata
}
\description{
A string list containing gross and net metadata.
}
\examples{
gross_net_metadata
}
\keyword{datasets}
