% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_maps.R
\name{load_aggregation_maps}
\alias{load_aggregation_maps}
\title{Load aggregation maps}
\usage{
load_aggregation_maps(
  path,
  aggregation_file_tabs = PFUAggDatabase::aggregation_file_tab_names,
  many_colname = PFUAggDatabase::aggregation_file_cols$many_colname,
  few_colname = PFUAggDatabase::aggregation_file_cols$few_colname
)
}
\arguments{
\item{path}{The path to an Excel file containing aggregation tables.}

\item{aggregation_file_tabs}{A list of tabs in the aggregation file.
The names are the programmatic names for the tabs.
The items are the actual names.
Default is \code{PFUAggDatabase::aggregation_file_tabs}.}

\item{many_colname}{The name for the many column in the aggregation tables.
Default is \code{PFUAggDatabase::aggregation_file_info$many_colname}.}

\item{few_colname}{The name for the few column in the aggregation tables.
Default is \code{PFUAggDatabase::aggregation_file_info$few_colname}.}
}
\value{
A named list of aggregation maps.
The names give the aggregation map.
The items are the aggregation maps themselves.
}
\description{
Load aggregation tables and create aggregation maps
for use in the workflow.
}
\details{
An aggregation table consists of a many column and a few column.
An aggregation map is a named list of items, where the names are the few
and the items are the many.
}
