% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_file_info}
\alias{output_file_info}
\title{Output file information}
\format{
A string list with 3 entries.
\describe{
\item{agg_eta_filename}{The name of the Excel file containing aggregate energy and exergy values and aggregate efficiencies.}
\item{agg_tabname}{The name of the tab in \code{agg_eta_filename} containing aggregated energy and exergy values.}
\item{eta_tabname}{The name of the tab in \code{agg_eta_filename} containing aggregated efficiencies.}
}
}
\usage{
output_file_info
}
\description{
A string list containing named names of files and tabs for outputs from this pipeline.
}
\examples{
output_file_info
}
\keyword{datasets}
