% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{pfu_aggregations}
\alias{pfu_aggregations}
\title{Aggregate to primary, final, and useful sums}
\usage{
pfu_aggregations(.psut_data)
}
\arguments{
\item{.psut_data}{A data frame in wide-by-matrices format containing PSUT matrices.}
}
\value{
A data frame of primary, final, and useful aggregations for each region.
}
\description{
Aggregate to primary, final, and useful sums
}
