% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pivot_agg_eta_wide_by_year}
\alias{pivot_agg_eta_wide_by_year}
\title{Pivot an aggregation or efficiency data frame wide by years}
\usage{
pivot_agg_eta_wide_by_year(
  .df,
  pivot_cols,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{.df}{The data frame to pivot.}

\item{pivot_cols}{The columns to pivot.}

\item{country, year}{See \code{IEATools::iea_cols}.}

\item{quantity, .values}{See \code{IEATools::template_cols}.}
}
\value{
A version of .df that is pivoted wide by years.
}
\description{
Converts a aggregation of efficiencies data frame (in tidy format)
to one that is wide by years.
}
