% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrIn_aggregations.R
\name{rename_prime_cols}
\alias{rename_prime_cols}
\title{Rename chopped matrices and replace full matrices}
\usage{
rename_prime_cols(
  .chopped_eccs,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  .prime = "_prime"
)
}
\arguments{
\item{.chopped_eccs}{A data frame of chopped energy conversion chain matrices, wide by matrices.}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed, S_units}{The names of input columns in \code{.psut_data}.
Default values are from \code{Recca::psut_cols}.}

\item{.prime}{The string suffix of the matrices representing the chopped ECC.
Default is "_prime".}
}
\value{
\code{.chopped_eccs} with modified columns and column names.
}
\description{
During the chopping process,
\code{\link[Recca:chop-doc]{Recca::chop_R()}} and \code{\link[Recca:chop-doc]{Recca::chop_Y()}}
create a data frame with both the original matrices
(in \code{matsindf} columns with standard RUVY names) and
ECC matrices representing chopped \strong{R} and \strong{Y} matrices
named with \code{.prime} suffixes.
This function deletes the original ECC matrices
and replaces with renamed \code{.prime} matrices.
For example, the column of \strong{R} matrices is deleted, and
the \strong{R_prime} column is renamed "R".
}
