% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopRY.R
\name{rename_prime_psut_columns}
\alias{rename_prime_psut_columns}
\title{Rename "_prime" columns}
\usage{
rename_prime_psut_columns(
  .psut_data,
  suffix = "_prime",
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices.}

\item{suffix}{The suffix for columns of aggregated matrices.
Default is "_prime".}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed}{The names of input columns in \code{.psut_data}.}
}
\value{
A version of \code{.psut_data} with original PSUT columns deleted and suffixed columns renamed to original names.
}
\description{
Deletes the original (un-prime) columns and renames the prime columns to those names
}
