% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPr_aggregations.R
\name{rename_suffixed_psut_columns}
\alias{rename_suffixed_psut_columns}
\title{Delete original and rename aggregated columns}
\usage{
rename_suffixed_psut_columns(
  .psut_data,
  suffix = Recca::aggregate_cols$aggregated_suffix,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices.}

\item{suffix}{The suffix for columns of aggregated matrices.
Default is \code{Recca::aggregate_cols$aggregated_suffix}.}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed, S_units}{The names of input columns in \code{.psut_data}.}
}
\value{
A version of \code{.psut_data} with original PSUT columns deleted and suffixed columns renamed to original names.
}
\description{
This is a helpful function that deletes original PSUT columns and
renames the aggregated columns to the original names.
}
\details{
Columns with names given by \code{R}, \code{U}, \code{V}, \code{Y}, \code{r_eiou}, \code{U_eiou}, \code{U_feed}, and \code{S_units} arguments.
The suffix is given by the \code{suffix} argument.

This function is a helper function and, thus, is not public.
}
