% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_cols}
\alias{sea_cols}
\title{Data frame column names}
\format{
A string list with 10 entries.
\describe{
\item{stage_colname}{The name of the metadata column that tells about the stage of the energy conversion chain. "Stage".}
\item{gross_net_colname}{The name of the metadata column that indicates gross or net energy. "Gross.Net"}
\item{e_product_colname}{The name of the metadata column that indicates energy products. "E.product".}
\item{e_sector_colname}{The name of the metadata column that indicates energy sectors. "Sector".}
\item{flow_colname}{The name of the metadata column that indicates IEA flows. "Flow".}
\item{agg_by_colname}{The name of the metadata column that indicates how things are aggregated. "Aggregation.by".}
\item{fd_sectors_colname}{The name of the metadata column that indicates final demand sectors. "Fd.sectors".}
\item{p_ind_comp_colname}{The name of the metadata column that indicates primary industries. "p_industries_colname".}
\item{p_ind_prefix_colname}{The name of the metadata column that indicates primary industry indices. "p_industry_prefixes".}
\item{ex_colname}{The name of the metadata column that indicates energy or exergy. "EX".}
}
}
\usage{
sea_cols
}
\description{
A string list containing names of column names and values for aggregation data frames.
}
\examples{
sea_cols
}
\keyword{datasets}
