% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopRY.R
\name{stack_chop_ECCs}
\alias{stack_chop_ECCs}
\title{Stack chopped ECC data frames}
\usage{
stack_chop_ECCs(
  PSUT_Re_all_Gr_all,
  PSUT_Re_all_Gr_all_Chop_Y,
  PSUT_Re_all_Gr_all_Chop_R,
  chop_mat = PFUAggDatabase::aggregation_df_cols$chopped_mat,
  Y_matname = "Y",
  R_matname = "R",
  chop_var = PFUAggDatabase::aggregation_df_cols$chop_var,
  product_sector = PFUAggDatabase::aggregation_df_cols$product_sector,
  none = "None"
)
}
\arguments{
\item{none}{}
}
\description{
Stack chopped ECC data frames
}
