% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrIn_aggregations.R
\name{stack_despecification_aggregations}
\alias{stack_despecification_aggregations}
\title{Stack all product and industry aggregations}
\usage{
stack_despecification_aggregations(
  specified_df,
  Ds_Pr = NULL,
  Ds_In = NULL,
  Ds_PrIn = NULL,
  product_aggregation = PFUAggDatabase::aggregation_df_cols$product_aggregation,
  industry_aggregation = PFUAggDatabase::aggregation_df_cols$industry_aggregation,
  specified = PFUAggDatabase::aggregation_df_cols$specified,
  despecified = PFUAggDatabase::aggregation_df_cols$despecified
)
}
\arguments{
\item{specified_df}{A data frame in which all rows and columns are specified.}

\item{Ds_Pr}{A data frame in which products are despecified.}

\item{Ds_In}{A data frame in which industries are despecified.}

\item{Ds_PrIn}{A data frame in which both products and industries are despecified.}

\item{product_aggregation, industry_aggregation, specified, despecified}{See \code{PFUAggDatabase::aggregation_df_cols}.}
}
\value{
A stacked data frame containing new metadata columns for product and industry specification.
}
\description{
All product and industry aggregations need to be present in the same data frame.
This function stacks them (with \code{dplyr::bind_rows})
and adds columns to identify the levels of product and industry aggregation.
}
