% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPr_aggregations.R
\name{stack_group_aggregations}
\alias{stack_group_aggregations}
\title{Stack grouped data frames}
\usage{
stack_group_aggregations(
  not_grouped,
  Gr_Pr = NULL,
  Gr_In = NULL,
  Gr_PrIn = NULL,
  product_aggregation = PFUAggDatabase::aggregation_df_cols$product_aggregation,
  industry_aggregation = PFUAggDatabase::aggregation_df_cols$industry_aggregation,
  ungrouped = PFUAggDatabase::aggregation_df_cols$ungrouped,
  grouped = PFUAggDatabase::aggregation_df_cols$grouped
)
}
\arguments{
\item{not_grouped}{The data frame with no product or industry grouping.}

\item{Gr_Pr}{The data frame with grouped products.}

\item{Gr_In}{The data frame with grouped industries.}

\item{Gr_PrIn}{The data frame with both grouped products and industries.}
}
\value{
A stacked set of data frames with different product and industry groupings
}
\description{
Matrices can be grouped by products and industries.
This function stacks those data frames with appropriate metadata columns
using \code{dplyr::bind_rows()}.
}
