% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_aggregations.R
\name{world_aggregation}
\alias{world_aggregation}
\title{Aggregate continents to world}
\usage{
world_aggregation(
  PSUT_Chop_all_Re_continents,
  years,
  world_aggregation_map,
  country = Recca::psut_cols$country,
  world = "World",
  year = Recca::psut_cols$year,
  many_colname = Recca::psut_cols$country,
  few_colname = world
)
}
\arguments{
\item{PSUT_Chop_all_Re_continents}{A data frame of PSUT matrices for continents.}

\item{years}{The years for which aggregation is desired.}

\item{world_aggregation_map}{The aggregation map to sum continents into the world.}

\item{country}{The name of the country column in the \code{PSUT_Chop_all_Re_continents} data frame.
Default is \code{Recca::psut_cols$country}.}

\item{world}{The name of the column of few things that will remain on output.
Default is "World".}

\item{year}{The name of the year column in the \code{PSUT} data frame.
Default is \code{Recca::psut_cols$year}.}

\item{many_colname}{The name of the column of many things that will be aggregated into continents.
Default is \code{Recca::psut_cols$country}.}

\item{few_colname}{The name of the column of few things that will remain on output.
Default is the same as the value of the \code{world} argument.}
}
\value{
A data frame in which continents are aggregated to world
according to the aggregation map
for the desired \code{years}.
}
\description{
This function is a wrapper for \code{Recca::region_aggregates()}
that also filters on \code{years},
thereby enabling parallel processing across years.
}
