% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{write_agg_etas_xlsx}
\alias{write_agg_etas_xlsx}
\title{Pivot and write aggregate efficiencies to an Excel file}
\usage{
write_agg_etas_xlsx(
  .agg_etas,
  path,
  aggs_tabname = PFUAggDatabase::output_file_info$agg_tabname,
  etas_tabname = PFUAggDatabase::output_file_info$eta_tabname,
  wide_by_year = TRUE,
  primary = IEATools::all_stages$primary,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  eta_pf_colname = PFUAggDatabase::efficiency_cols$eta_pf,
  eta_fu_colname = PFUAggDatabase::efficiency_cols$eta_fu,
  eta_pu_colname = PFUAggDatabase::efficiency_cols$eta_pu,
  year = IEATools::iea_cols$year,
  country = IEATools::iea_cols$country,
  quantity = IEATools::template_cols$quantity,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{.agg_etas}{A data frame created by the \code{eta_Re_all_St_pfu} target.}

\item{path}{The path where the Excel file will be saved.}

\item{aggs_tabname, etas_tabname}{See \code{PFUAggDatabase::output_file_info}.}

\item{wide_by_year}{If \code{TRUE} (the default), the incoming data frame will be pivoted to be wide by years.
If \code{FALSE}, data will be unchanged.}

\item{primary, final, useful}{See \code{IEATools::all_stages}.}

\item{eta_pf_colname, eta_fu_colname, eta_pu_colname}{See \code{PFUAggDatabase::efficiency_cols}.}

\item{year, country}{See \code{IEATools::iea_cols}.}

\item{quantity, .values}{See \code{IEATools::template_cols}.}
}
\value{
\code{TRUE} if the file was written successfully.
}
\description{
The incoming data frame is expected to contain
a \code{year} column as well as
\code{primary}, \code{final}, \code{useful}, \code{eta_pf}, \code{eta_fu}, and \code{eta_pu} columns.
}
