% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{write_eta_pfu_xlsx}
\alias{write_eta_pfu_xlsx}
\title{Write an Excel file of efficiency results}
\usage{
write_eta_pfu_xlsx(.eta_pfu, release = FALSE, path = NULL, overwrite = FALSE)
}
\arguments{
\item{.eta_pfu}{A data frame of pf, fu, and pu efficiencies.}

\item{release}{A boolean that tells whether to write the file. Default is \code{FALSE}.}

\item{overwrite}{A boolean that tells whether to overwrite an existing file at \code{path}. Default is \code{FALSE}.}

\item{directory}{A folder for the output.
A subfolder and file name are calculated automatically,
following the pins naming convention.}
}
\value{
If a file is written, the value of \code{path}.
If no file is written
(e.g., because \code{overwrite = FALSE} and the file already exists),
\code{character(0)}.
}
\description{
Writes an Excel file of efficiency results.
}
