% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agg_metadata}
\alias{agg_metadata}
\title{Aggregation metadata}
\format{
A string list with 5 entries.
\describe{
\item{total_value}{The string indicating aggregation totals. "Total".}
\item{all_value}{The string indicating all of something is aggregated. "All".}
\item{product_value}{The string indicating that products are aggregated. "Product".}
\item{sector_value}{The string indicating that sectors are aggregated. "Sector".}
\item{flow_value}{The string indicating that flows are aggregated. "Flow".}
}
}
\usage{
agg_metadata
}
\description{
A string list containing aggregation metadata.
}
\examples{
agg_metadata
}
\keyword{datasets}
