% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_file_cols}
\alias{aggregation_file_cols}
\title{Aggregation file column names}
\format{
A string list with 2 entries.
\describe{
\item{many_colname}{The string name of the many column. "Many".}
\item{few_colname}{The string name of the few column. "Few".}
}
}
\usage{
aggregation_file_cols
}
\description{
A string list containing aggregation file column names.
}
\examples{
aggregation_file_cols
}
\keyword{datasets}
