% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_file_tab_names}
\alias{aggregation_file_tab_names}
\title{Aggregation file tabs}
\format{
A string list with 2 entries.
\describe{
\item{continent_aggregation_tab}{The name of the continent aggregation tab.}
\item{world_aggregation_tab}{The name of the world aggregation tab.}
\item{exiobase_region_aggregation_tab}{The name of the Exiobase region aggregation tab.}
\item{eu.product_aggregation_tab}{The name of the useful product aggregation tab.}
\item{ef.product_aggregation_tab}{The name of the final product aggregation tab.}
\item{destination_aggregation_tab}{The name of the destination aggregation tab.}
}
}
\usage{
aggregation_file_tab_names
}
\description{
A string list containing names of tabs for an Excel aggregation file.
}
\details{
The aggregation file should have several tabs, one for each type of aggregation.
}
\examples{
aggregation_file_tab_names
}
\keyword{datasets}
