% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_pipelines.R
\name{cache_target}
\alias{cache_target}
\title{Last target in the pipeline}
\usage{
cache_target(agg_sym_dependency, eta_sym_dependency)
}
\value{
A logical saying whether the saving operation was successful.
If \code{release = FALSE}, \code{FALSE} is returned.
}
\description{
Provides a final target to store the cache of the pipeline.
}
