% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_finaldemand_aggregates}
\alias{calculate_finaldemand_aggregates}
\title{Calculate final demand aggregates for PSUT data}
\usage{
calculate_finaldemand_aggregates(
  .psut_data,
  countries,
  years,
  fd_sectors,
  pattern_type = "leading"
)
}
\arguments{
\item{.psut_data}{The data for which final demand aggregates are to be calculated.}

\item{countries}{The countries for which final demand aggregates are to be calculated.}

\item{years}{The years for which final demand aggregates are to be calculated.}

\item{pattern_type}{The type of matching to be used for final demand sectors names.
Default is "leading".}

\item{fd_industries}{A string vector of sectors that count as "final demand".}
}
\value{
A version of \code{.psut_data} with additional column for final demand aggregate data.
}
\description{
This function routes to \code{Recca::finaldemand_aggregates()}.
}
