% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_aggregations.R
\name{calculate_footprint_aggregations}
\alias{calculate_footprint_aggregations}
\title{Calculate footprint aggregates}
\usage{
calculate_footprint_aggregations(
  .psut_data,
  countries,
  years,
  p_industries,
  fd_sectors
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT matrices. It should be wide by matrices.}

\item{countries}{The countries to analyze.}

\item{years}{The years to analyze.}

\item{p_industries}{Industries that count for primary energy aggregates.}

\item{fd_sectors}{Final demand sectors that count for final demand aggregates.}
}
\value{
A data frame of footprint aggregates.
}
\description{
This function uses \code{Recca::footprint_aggregates()} internally
to calculate footprint aggregates.
}
