% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calculate_pfd_efficiencies}
\alias{calculate_pfd_efficiencies}
\title{Calculate primary to final demand efficiency}
\usage{
calculate_pfd_efficiencies(
  .agg_pfd_data,
  countries,
  years,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  eta_pfd = Recca::efficiency_cols$eta_pfd
)
}
\arguments{
\item{.agg_pfd_data}{A data frame of primary and final demand aggregates.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{ex_p, ex_fd}{The names of columns containing primary and final demand (respectively) aggregates.
See \code{Recca::aggregate_cols} for defaults.}

\item{eta_pfd}{The name of the primary-to-final demand efficiency column.
Default is \code{Recca::efficiency_cols$eta_pfd}.}
}
\value{
A data frame of primary-to-final-demand efficiencies, both net and gross.
}
\description{
This function calculates primary to final demand efficiencies in a data frame of
primary and final demand aggregates (\code{.pfd_data}).
Columns of energy conversion chain matrices are removed.
}
