% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calculate_pfu_efficiencies}
\alias{calculate_pfu_efficiencies}
\title{Pivot data frame to calculate PFU efficiencies}
\usage{
calculate_pfu_efficiencies(
  .eta_pfd_data,
  countries,
  years,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  last_stage = Recca::psut_cols$last_stage,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu,
  eta_pfd = Recca::efficiency_cols$eta_pfd,
  final = "Final",
  useful = "Useful"
)
}
\arguments{
\item{.eta_pfd_data}{A data frame produced by \code{calculate_pfd_efficiencies()}.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{ex_p, ex_fd}{Names of primary and final demand aggregate columns (respectively).
Defaults are from \code{Recca::aggregate_cols}.}

\item{last_stage}{The name of the last stage column. Default is \code{Recca::psut_cols$last_stage}.}

\item{eta_pf, eta_fu, eta_pu}{The names of efficiency columns: primary-to-final, final-to-useful, and primary-to-useful, respectively.
Defaults from \code{Recca::efficiency_cols}.}

\item{eta_pfd}{The name of a column containing primary-to-final demand efficiencies. Default is \code{Recca::efficiency_cols$eta_pfd}.}

\item{final, useful}{Entries in the \code{last_stage} column.
Defaults are "Final" and "Useful", respectively.}
}
\value{
A data frame with ECC stage efficiencies.
}
\description{
This function pivots the data frame produced by \code{calculate_pfd_efficiencies()}
to obtain primary-final, final-useful, and primary-useful efficiencies.
}
