% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calculate_pfu_efficiencies}
\alias{calculate_pfu_efficiencies}
\title{Pivot data frame to calculate PFU efficiencies}
\usage{
calculate_pfu_efficiencies(
  .eta_pfd_data,
  countries,
  years,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  last_stage = Recca::psut_cols$last_stage,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_pu = Recca::efficiency_cols$eta_pu,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pfd = Recca::efficiency_cols$eta_pfd,
  final = "Final",
  useful = "Useful"
)
}
\arguments{
\item{.eta_pfd_data}{A data frame produced by \code{calculate_pfd_efficiencies()}.}
}
\value{
A data frame with ECC stage efficiencies.
}
\description{
This function pivots the data frame produced by \code{calculate_pfd_efficiencies()}
to obtain primary-final, final-useful, and primary-useful efficiencies.
}
