% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_primary_aggregates}
\alias{calculate_primary_aggregates}
\title{Calculate primary aggregates for PSUT data}
\usage{
calculate_primary_aggregates(
  .psut_data,
  countries,
  years,
  p_industries,
  piece = "noun",
  notation = list(RCLabels::bracket_notation, RCLabels::arrow_notation),
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list
)
}
\arguments{
\item{.psut_data}{The data for which primary aggregates are to be calculated.}

\item{countries}{The countries for which primary aggregates are to be calculated.}

\item{years}{The years for which primary aggregates are to be calculated.}

\item{p_industries}{A string vector of industries that count as "primary".}

\item{pattern_type}{The type of matching to be used for primary industry names.
Default is "leading".}
}
\value{
A version of \code{.psut_data} with additional column for primary aggregate data.
}
\description{
This function routes to \code{Recca::primary_aggregates()}.
}
