% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficiency_cols}
\alias{efficiency_cols}
\title{PFUAggDatabase efficiency data frame column names}
\format{
A string list with 3 entries.
\describe{
\item{eta_pf}{The name of a column containing primary-to-final efficiencies.}
\item{eta_fu}{The name of a column containing final-to-useful efficiencies.}
\item{eta_pu}{The name of a column containing primary-to-useful efficiencies.}
}
}
\usage{
efficiency_cols
}
\description{
A string list containing named names of columns in the efficiency data frame of the
PFUAggDatabase.
}
\examples{
efficiency_cols
}
\keyword{datasets}
