% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{efficiency_pipeline}
\alias{efficiency_pipeline}
\title{Calculate PFU aggregations and efficiencies}
\usage{
efficiency_pipeline(
  .psut_data,
  countries,
  years,
  p_industries,
  fd_sectors,
  piece = "noun",
  notation = list(RCLabels::bracket_notation, RCLabels::arrow_notation),
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units,
  country = Recca::psut_cols$country,
  year = Recca::psut_cols$year,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  gross_net = Recca::efficiency_cols$gross_net,
  last_stage = Recca::psut_cols$last_stage,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_f = IEATools::aggregate_cols$aggregate_final,
  ex_u = IEATools::aggregate_cols$aggregate_useful,
  ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
  ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu
)
}
\arguments{
\item{.psut_data}{A data frame of PSUT data or a slice (row) of the data frame.}
}
\value{
A data frame of metadata columns and efficiencies
}
\description{
Calculate PFU aggregations and efficiencies
}
