% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_pipelines.R
\name{get_one_middle_pipeline}
\alias{get_one_middle_pipeline}
\title{A pipeline for one PSUT release}
\usage{
get_one_middle_pipeline(pr, aggregation_maps_tar_str, continents_tar_str)
}
\arguments{
\item{pr}{The \code{PFUDatabase} pipeline release for which this pipeline
segment will be constructed.}
}
\value{
A list of targets and cache dependencies for the incoming PSUT release (\code{pr}).
The list of targets is named "targets".
The list of cache dependencies is named "cache_dependencies".
}
\description{
\code{PFUAggDatabase} works with any number of PSUT releases from \code{PFUDatabase}.
This function creates a portion of a \code{targets} pipeline
for \emph{one} of those releases.
}
\details{
This function also creates a vector of cache dependencies,
targets that must be completed before the cache can be saved.
}
