% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_plans.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Create a targets workflow}
\usage{
get_pipeline()
}
\value{
A list of \code{tar_target}s to be executed in a workflow.
}
\description{
Arguments to this function specify the details of a targets workflow to be executed.
}
