% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pipelines.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Create a targets workflow}
\usage{
get_pipeline(
  countries = "all",
  years = "all",
  do_chops,
  psut_release,
  phi_vecs_release,
  psut_without_neu_release,
  aggregation_maps_path,
  pipeline_releases_folder,
  pipeline_caches_folder,
  release = FALSE
)
}
\arguments{
\item{countries}{A string vector of 3-letter country codes.
Default is "all", meaning all available countries should be analyzed.}

\item{years}{A numeric vector of years to be analyzed.
Default is "all", meaning all available years should be analyzed.}

\item{do_chops}{A boolean that tells whether to perform the R and Y chops.}

\item{psut_release}{The release we'll use from \code{pipeline_releases_folder}.
See details.}

\item{phi_vecs_release}{The release of the phi_vecs pin that we'll use.}
\item{psut_without_neu_release}{The release we'll use from \code{pipeline_releases_folder}.
See details.}

\item{aggregation_maps_path}{The path to the Excel file of aggregation maps.}

\item{pipeline_releases_folder}{The path to a folder where releases of output targets are pinned.}

\item{pipeline_caches_folder}{The path to a folder where releases of pipeline caches are stored.}

\item{release}{Boolean that tells whether to do a release of the results.
Default is \code{FALSE}.}
}
\value{
A list of \code{tar_target}s to be executed in a workflow.
}
\description{
This is a target factory whose arguments
specify the details of a targets workflow to be constructed
}
