% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_file_info}
\alias{output_file_info}
\title{Output file information}
\format{
A string list with 3 entries.
\describe{
\item{agg_eta_filename}{A string value of the output file name. "AggregateEfficiencyResults.xlsx".}
\item{agg_tabname}{The string name of the aggregates tab. "Aggregates".}
\item{eta_tabname}{The string name of the efficiency tab. "etas".}
}
}
\usage{
output_file_info
}
\description{
A string list containing the filename and tabs for output files.
}
\examples{
output_file_info
}
\keyword{datasets}
