% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_aggregations.R
\name{region_pipeline}
\alias{region_pipeline}
\title{Aggregation regions}
\usage{
region_pipeline(
  .psut_data,
  continent_aggregation_map,
  world_aggregation_map,
  country = Recca::psut_cols$country,
  year = Recca::psut_cols$year,
  continent = "Continent",
  world = "World"
)
}
\arguments{
\item{.psut_data}{}
}
\value{
A data frame that includes new "Country"s for
continents and the World.
}
\description{
The database can benefit from continent and World aggregations.
This function bundles those aggregations into a single function.
}
\details{
All regional aggregations have names that are 5 characters or longer.
}
