% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_subpipelines.R
\name{setup_targets}
\alias{setup_targets}
\title{Create setup targets for PFUAggDatabase}
\usage{
setup_targets(
  countries,
  years,
  aggregation_maps_path,
  pipeline_releases_folder,
  release,
  aggregation_maps_tar_str,
  continents_tar_str
)
}
\arguments{
\item{countries}{A string vector of 3-letter country codes.
Default is "all", meaning all available countries should be analyzed.}

\item{years}{A numeric vector of years to be analyzed.
Default is "all", meaning all available years should be analyzed.}

\item{aggregation_maps_path}{The path to the Excel file of aggregation maps.}

\item{pipeline_releases_folder}{The path to a folder where releases of output targets are pinned.}

\item{release}{Boolean that tells whether to do a release of the results.
Default is \code{FALSE}.}

\item{psut_releases}{The releases we'll use from \code{pipeline_releases_folder}.
See details.}
}
\value{
A list of initial targets.
}
\description{
There are several initial targets that are invariant
to the number of psut_releases that need to be analyzed
by this workflow.
This function creates those initial targets.
}
\details{
Some targets created by this function
have names that are known and assumed by \code{get_one_middle_pipeline()}.
This function and \code{get_one_middle_pipeline()} should be considered
as a pair of functions that work together.
}
