% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPr_aggregations.R
\name{stack_PrIn_aggregations}
\alias{stack_PrIn_aggregations}
\title{Stack all product and industry aggregations}
\usage{
stack_PrIn_aggregations(
  PSUT_Re_all,
  PSUT_Re_all_Pr_despec_In_despec,
  PSUT_Re_all_Pr_group,
  PSUT_Re_all_In_group,
  PSUT_Re_all_Pr_group_In_group,
  product_aggregation = PFUAggDatabase::aggregation_df_cols$product_aggregation,
  industry_aggregation = PFUAggDatabase::aggregation_df_cols$industry_aggregation,
  specified = PFUAggDatabase::aggregation_df_cols$specified,
  despecified = PFUAggDatabase::aggregation_df_cols$despecified,
  grouped = PFUAggDatabase::aggregation_df_cols$grouped
)
}
\arguments{
\item{PSUT_Re_all, PSUT_Re_all_Pr_despec_In_despec, PSUT_Re_all_Pr_group, PSUT_Re_all_In_group, PSUT_Re_all_Pr_group_In_group}{Data frames to be stacked.}

\item{product_aggregation, industry_aggregation, specified, despecified, grouped}{See \code{PFUAggDatabase::aggregation_df_cols}.}
}
\value{
A stacked data frame containing new metadata columns for product and industry aggregations.
}
\description{
All product and industry aggregations need to be present in the same data frame.
This function stacks them (with \code{dplyr::bind_rows})
and adds columns to identify the levels of product and industry aggregation.
}
